Move methods used for dependency graph into DependencyGraphGenerator
This alligns to the extending AS poet guide, in particular, it separates
the generator from the blueprint.
