+++ src/main/kotlin/com/google/androidstudiopoet/AndroidStudioPoet.kt
@@ -165,7 +165,6 @@ class AndroidStudioPoet(private val modulesGenerator: SourceModuleGenerator, pri
         val timeSpent = measureTimeMillis {
             projectBluePrint = ProjectBlueprint(projectConfig)
             modulesGenerator.generate(projectBluePrint!!)
-            projectBluePrint!!.saveDependencies()
             if (projectBluePrint!!.hasCircularDependencies()) {
                 println("WARNING: there are circular dependencies")
             }
